refactor: split cli parsing helpers
